+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Unreleased
 * #502 Added support for system property in query parameters
 * #503 `request~toJSON` no longer converts `url` to string
+* Made sure all Script instances have an id
 
 #### v3.0.3 (November 8, 2017)
 * :arrow_up: Updated dependencies.