+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * #502 Added support for system property in query parameters
 * #503 `request~toJSON` no longer converts `url` to string
 * Made sure all Script instances have an id
+* :bug: `auth` is now included in `collection.toJSON`
 
 #### v3.0.3 (November 8, 2017)
 * :arrow_up: Updated dependencies.